step4/django-chat/requirements.txt
Django==5.2.6
djangorestframework==3.16.1
channels==4.3.1
channels-redis==4.3.0
redis==6.4.0
asgiref==3.9.1